refactor: modulize Errocheck test
